fix some valgrind errors with ft_calloc
